Move setup into setupTests
